rename store to configureStore
