+++ app/utils/tests/injectReducer.test.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme'
 import React from 'react'
 import identity from 'lodash/identity'
 
-import configureStore from '../../store'
+import configureStore from '../../configureStore'
 import injectReducer from '../injectReducer'
 import * as reducerInjectors from '../reducerInjectors'
